+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # orchid-core
 
+## 0.3.1
+
+### Patch Changes
+
+- Accept string, number, Date in date time columns
+
 ## 0.3.0
 
 ### Minor Changes