+++ packages/qb/myqb/CHANGELOG.md
@@ -1,5 +1,12 @@
 # myqb
 
+## 0.2.2
+
+### Patch Changes
+
+- Updated dependencies
+  - orchid-core@0.3.1
+
 ## 0.2.1
 
 ### Patch Changes