+++ packages/test-factory/CHANGELOG.md
@@ -1,5 +1,15 @@
 # orchid-orm-test-factory
 
+## 0.3.44
+
+### Patch Changes
+
+- Updated dependencies
+  - pqb@0.11.2
+  - orchid-core@0.3.1
+  - orchid-orm@1.8.1
+  - orchid-orm-schema-to-zod@0.3.40
+
 ## 0.3.43
 
 ### Patch Changes